[refactor] FCM 서비스 패키지 리팩터링 및 에러 핸들링
