+++ src/main/java/com/example/momogum/apiPayLoad/exception/handler/TokenHandler.java
@@ -1,0 +1,10 @@
+package com.example.momogum.apiPayLoad.exception.handler;
+
+import com.example.momogum.apiPayLoad.code.BaseErrorCode;
+import com.example.momogum.apiPayLoad.exception.GeneralException;
+
+public class TokenHandler extends GeneralException {
+    public TokenHandler(BaseErrorCode code) {
+        super(code);
+    }
+}
