+++ src/main/java/com/example/momogum/util/FirebaseCloudMessageUtil.java
@@ -1,14 +1,14 @@
-package com.example.momogum.fcm;
+package com.example.momogum.util;
 
 import com.example.momogum.apiPayLoad.code.status.ErrorStatus;
+import com.example.momogum.apiPayLoad.exception.handler.TokenHandler;
 import com.example.momogum.apiPayLoad.exception.handler.UserEntityHandler;
 import com.example.momogum.domain.UserEntity;
+import com.example.momogum.web.dto.FcmMessage;
 import com.example.momogum.repository.userEntityRepo.UserEntityRepository;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.google.auth.oauth2.GoogleCredentials;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,6 +17,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
@@ -27,36 +28,36 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class FirebaseCloudMessageService {
+public class FirebaseCloudMessageUtil {
 
-    private final String API_URL = "https://fcm.googleapis.com/v1/projects/momogum-d9bfb/messages:send";
+    @Value("${fcm.api-url}")
+    private String API_URL;
+
     private final ObjectMapper objectMapper;
     private final UserEntityRepository userEntityRepository;
 
-    public void sendMessageTo(Long userId, String title, String body) throws IOException {
 
-        log.info("------------푸시알림 시작------------");
+    /**
+     * FCM 푸시 알림을 전송하는 메서드 입니다
+     *
+     * 알림을 전송하는 대상의 PK와 알림 제목, 내용을 넣어서 호출해주세요
+     * */
+    public void sendMessageTo(Long userId, String title, String body) throws IOException {
 
         UserEntity byId = userEntityRepository.findById(userId)
                 .orElseThrow(() -> new UserEntityHandler(ErrorStatus.MEMBER_NOT_FOUND));
 
         if (byId.getFcmToken() == null || byId.getFcmToken().isEmpty()) {
-            throw new IllegalStateException("FCM 토큰이 없습니다.");
+            throw new TokenHandler(ErrorStatus.FCM_TOKEN_NOT_FOUND);
         }
 
-        log.info("토큰을 발견했습니다: " + byId.getFcmToken());
-
         String message = makeMessage(byId.getFcmToken(), title, body);
-
-        log.info("FCM에 전송할 메세지를 만들었습니다 :" + message);
 
         OkHttpClient client = new OkHttpClient();
         RequestBody requestBody = RequestBody.create(
                 MediaType.parse("application/json"),
                 message
         );
-
-        log.info("requestBody를 제작하였습니다: " + requestBody);
 
         Request request = new Request.Builder()
                 .url(API_URL)
@@ -65,17 +66,16 @@ public class FirebaseCloudMessageService {
                 .addHeader(HttpHeaders.CONTENT_TYPE, "application/json")
                 .build();
 
-        log.info("AccessToken 입니다: " + getAccessToken());
-
         try (Response response = client.newCall(request).execute()) {
             String responseBody = response.body().string();
             if (!response.isSuccessful()) {
-                System.out.println("FCM Error Response: " + responseBody);  // 에러 응답 로깅
+                System.out.println("FCM Error Response: " + responseBody);
                 throw new IOException("FCM 메시지 전송 실패: " + response.code());
             }
-            System.out.println("FCM Success Response: " + responseBody);  // 성공 응답 로깅
+            System.out.println("FCM Success Response: " + responseBody);
         }
     }
+
     private String makeMessage(String token, String title, String body) throws JsonParseException, JsonProcessingException {
         FcmMessage fcmMessage = FcmMessage.builder()
                 .message(FcmMessage.Message.builder()
@@ -93,17 +93,11 @@ public class FirebaseCloudMessageService {
     private String getAccessToken() throws IOException {
         String firebaseConfigPath = "firebase/firebase_service_key.json";
 
-        ClassPathResource resource = new ClassPathResource(firebaseConfigPath);
-        log.info("Firebase 설정 파일 존재 여부: " + resource.exists());
-        log.info("Firebase 설정 파일 경로: " + resource.getPath());
-
         GoogleCredentials googleCredentials = GoogleCredentials
                 .fromStream(new ClassPathResource(firebaseConfigPath).getInputStream())
                 .createScoped(List.of("https://www.googleapis.com/auth/cloud-platform"));
 
         googleCredentials.refreshIfExpired();
-
-        log.info("AccessToken 입니다: " + googleCredentials.getAccessToken().getTokenValue());
         return googleCredentials.getAccessToken().getTokenValue();
     }
 }
