+++ src/main/java/com/example/momogum/web/dto/FCMRequestDTO.java
@@ -1,11 +1,11 @@
-package com.example.momogum.fcm;
+package com.example.momogum.web.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class RequestDTO {
+public class FCMRequestDTO {
     private Long userId;
     private String title;
     private String body;
